+++ lib/OutputStreams/IndexOutputStream.h
@@ -65,7 +65,7 @@ class IndexOutputStream : public BaseOutputStream
                       const RealMatrix&    sourceMatrix,
                       const IndexMatrix&   sensorMask,
                       const ReduceOperator reduceOp,
-                      float*               bufferToReuse = nullptr);
+                      double*               bufferToReuse = nullptr);
 
     /// Copy constructor not allowed.
     IndexOutputStream(const IndexOutputStream&) = delete;
